added OneWay wall, bug fixes & started lvl 3&4
OneWayWall lets the player pass throught it from the one side, but
blocks it from the other side. Also added a gradient fade shader for the
graphics of the OneWayWall
Fixed a bug where the direction would be zero when we check our next
direction.
Finished lvl2 & Started with lvl 3 & 4