Fixed corner collision bug. Added onewaytriggercollider.
fixed a bug where there would be no collision in a corner due to
collision rigidbody correction overshooting. I now use raycasting to
detect corner collisions and collision detection for other collisision.
Added a onewaytriggercollider to know which way the player is coming
from and when to activate the trigger.